+++ tests/test_account.py
@@ -261,3 +261,28 @@ class TestAccount(TestCase):
             method='DELETE',
             params=None, url='https://api.toshl.com/accounts/1')
         assert response.json() == ''
+
+    @mock.patch('toshl.client.requests.request')
+    def test_move_account_successful(self, mock_request):
+        json_payload = {
+            'position': 2
+        }
+
+        mock_response = mock.Mock()
+        mock_response.json.return_value = ''
+        mock_response.status_code = 204
+        mock_request.return_value = mock_response
+
+        client = ToshlClient('abcd1234')
+        account = Account(client)
+        response = account.move('1', 2)
+
+        mock_request.assert_called_once_with(
+            headers={
+                'Authorization': 'Bearer abcd1234',
+                'Content-Type': 'application/json'
+            },
+            method='POST',
+            params=None, url='https://api.toshl.com/accounts/1/move',
+            json=json_payload)
+        assert response.json() == ''